test(d1): consolidate regression suites and unify documentation
