--- experiments/d1/manifests/licenses.md
@@ -1,21 +1,0 @@
-# D1 P0 data and teacher license record
-
-This file records upstream license locations for reproducibility. It is not legal advice and does not replace the
-upstream terms.
-
-## COCO 2017
-
-- Dataset home and terms: https://cocodataset.org/#termsofuse
-- Download source: http://images.cocodataset.org/
-- COCO does not grant a single blanket license for all source images. Each image remains subject to its original Flickr
-  license. The official COCO terms and image-level license metadata must be followed.
-- The generated YOLO detection labels are obtained from the Ultralytics COCO 2017 labels archive and are used only as a
-  representation of the canonical COCO annotations.
-
-## DINOv3 ViT-S/16
-
-- Model: https://www.modelscope.cn/models/facebook/dinov3-vits16-pretrain-lvd1689m
-- Upstream project: https://github.com/facebookresearch/dinov3
-- License: https://github.com/facebookresearch/dinov3/blob/main/LICENSE.md
-- DINOv3 code and model weights are distributed under the DINOv3 License. The license copy downloaded with the model
-  and its SHA256 are retained in the external experiment workspace and cloud backup.
