+++ tests/test_d1_cache.py
@@ -1,27 +1,300 @@
-"""Production NPY reader, provenance, pinned-memory alias, and resume regressions."""
+"""Safetensors and NPY integrity, recovery, and cached tensor consumption."""
 
 from __future__ import annotations
 
-import copy
 import json
+from pathlib import Path
+import pytest
+import torch
+from ultralytics.nn.foundation import cache as cache_module
+from ultralytics.nn.foundation.cache import (
+    FeatureCacheReader,
+    FeatureCacheWriter,
+    build_cache_key,
+    compare_feature_caches,
+    sha256_bytes,
+    verify_feature_cache,
+)
+import copy
 import os
 import pickle
 import shutil
-from pathlib import Path
 from types import SimpleNamespace
-
 import cv2
 import numpy as np
-import pytest
-import torch
 from torch.utils.data._utils.pin_memory import pin_memory
-
 from ultralytics.data.d1_cache import D1FeatureCacheDataset, D1TrainingBatch
 from ultralytics.models.yolo.detect.foundation_train import D1FoundationDetectionTrainer
 from ultralytics.models.yolo.detect.train import DetectionTrainer
-from ultralytics.nn.foundation.cache import FeatureCacheReader, FeatureCacheWriter, sha256_file
+from ultralytics.nn.foundation.cache import sha256_file
 from ultralytics.nn.foundation.npy_cache import NpyFeatureCacheReader, open_feature_cache, validate_npy_evidence
 from ultralytics.utils import DEFAULT_CFG_DICT
+
+
+def contract(**updates):
+    value = {
+        "schema_version": "d1-cache-v1",
+        "model_id": "local/test-teacher",
+        "teacher_weights_sha256": "a" * 64,
+        "preprocessing_sha256": "b" * 64,
+        "output_layers": [4, 8, 12],
+        "feature_names": ["block4", "block8", "block12"],
+        "dtype": "float16",
+        "expected_shape": [2, 2, 2],
+    }
+    value.update(updates)
+    return value
+
+
+def features(value):
+    return {
+        name: torch.full((2, 2, 2), float(value + index), dtype=torch.float32)
+        for index, name in enumerate(("block4", "block8", "block12"))
+    }
+
+
+def image_sha(value):
+    return sha256_bytes(f"image-{value}".encode())
+
+
+def add_sample(writer, value):
+    return writer.add(
+        sample_id=f"train2017/{value:012d}",
+        split="train2017",
+        image_path=f"images/train2017/{value:012d}.jpg",
+        image_sha256=image_sha(value),
+        features=features(value),
+    )
+
+
+def build_cache(path: Path, values=(1, 2)):
+    with FeatureCacheWriter(path, split="train2017", contract=contract(), target_shard_bytes=60) as writer:
+        for value in values:
+            assert add_sample(writer, value)
+
+
+def test_cache_key_changes_with_every_content_identity_field():
+    base = {
+        "image_sha256": "1" * 64,
+        "preprocessing_sha256": "2" * 64,
+        "teacher_weights_sha256": "3" * 64,
+        "output_layers": (4, 8, 12),
+        "dtype": "float16",
+    }
+    baseline = build_cache_key(**base)
+    for name, changed in (
+        ("image_sha256", "4" * 64),
+        ("preprocessing_sha256", "5" * 64),
+        ("teacher_weights_sha256", "6" * 64),
+        ("output_layers", (4, 12)),
+    ):
+        candidate = dict(base)
+        candidate[name] = changed
+        assert build_cache_key(**candidate) != baseline
+
+
+@pytest.mark.parametrize(
+    "updates, exception",
+    [
+        ({"dtype": "float32"}, ValueError),
+        ({"output_layers": []}, ValueError),
+        ({"output_layers": [8, 4]}, ValueError),
+        ({"feature_names": ["block4"]}, ValueError),
+        ({"expected_shape": [2, 2]}, ValueError),
+        ({"teacher_weights_sha256": "bad"}, ValueError),
+    ],
+)
+def test_invalid_cache_contract_fails_fast(tmp_path, updates, exception):
+    with pytest.raises(exception):
+        FeatureCacheWriter(tmp_path, split="train2017", contract=contract(**updates))
+
+
+def test_sharded_roundtrip_index_and_full_verification(tmp_path):
+    build_cache(tmp_path)
+
+    report = verify_feature_cache(tmp_path)
+    reader = FeatureCacheReader(tmp_path)
+    index = json.loads((tmp_path / "index.json").read_text(encoding="utf-8"))
+
+    assert report["sample_count"] == 2
+    assert report["shard_count"] == 2
+    assert report["tensor_count"] == 6
+    assert report["part_files"] == []
+    assert index["split_counts"] == {"train2017": 2}
+    assert index["content_sha256"] == report["content_sha256"]
+    assert all(not Path(record["image_path"]).is_absolute() for record in reader.records.values())
+    loaded = reader.get("train2017/000000000001")
+    assert tuple(loaded) == ("block4", "block8", "block12")
+    assert loaded["block4"].dtype == torch.float16
+    assert torch.all(loaded["block4"] == 1)
+    assert torch.all(loaded["block12"] == 3)
+
+
+def test_resume_skips_committed_sample_and_appends_next_shard(tmp_path):
+    build_cache(tmp_path, values=(1,))
+    writer = FeatureCacheWriter(tmp_path, split="train2017", contract=contract(), target_shard_bytes=60)
+
+    assert add_sample(writer, 1) is False
+    assert add_sample(writer, 2) is True
+    writer.close()
+
+    report = verify_feature_cache(tmp_path)
+    assert report["sample_count"] == 2
+    assert report["shard_count"] == 2
+
+
+def test_rank_shard_prefix_is_unique_and_resumes(tmp_path):
+    writer = FeatureCacheWriter(
+        tmp_path,
+        split="train2017",
+        contract=contract(),
+        target_shard_bytes=60,
+        shard_prefix="train2017-r03",
+    )
+    assert add_sample(writer, 1)
+    writer.close()
+    resumed = FeatureCacheWriter(
+        tmp_path,
+        split="train2017",
+        contract=contract(),
+        target_shard_bytes=60,
+        shard_prefix="train2017-r03",
+    )
+    assert add_sample(resumed, 1) is False
+    assert add_sample(resumed, 2)
+    resumed.close()
+
+    assert [path.name for path in sorted(tmp_path.glob("*.safetensors"))] == [
+        "train2017-r03-00000.safetensors",
+        "train2017-r03-00001.safetensors",
+    ]
+    assert verify_feature_cache(tmp_path)["sample_count"] == 2
+
+
+@pytest.mark.parametrize("prefix", ("rank00", "../train2017-r00", "train2017/r00", ""))
+def test_invalid_shard_prefix_fails_fast(tmp_path, prefix):
+    with pytest.raises(ValueError, match="shard_prefix"):
+        FeatureCacheWriter(tmp_path, split="train2017", contract=contract(), shard_prefix=prefix)
+
+
+def test_missing_index_is_rebuilt_from_committed_shard_headers(tmp_path):
+    build_cache(tmp_path)
+    expected = verify_feature_cache(tmp_path)["content_sha256"]
+    (tmp_path / "index.json").unlink()
+    (tmp_path / "samples.jsonl").unlink()
+
+    writer = FeatureCacheWriter(tmp_path, split="train2017", contract=contract(), target_shard_bytes=60)
+    writer.close()
+
+    assert verify_feature_cache(tmp_path)["content_sha256"] == expected
+
+
+def test_two_independent_builds_have_identical_content(tmp_path):
+    first, second = tmp_path / "first", tmp_path / "second"
+    build_cache(first)
+    build_cache(second)
+
+    comparison = compare_feature_caches(first, second)
+
+    assert comparison["identical"] is True
+    assert comparison["sample_count"] == 2
+    assert comparison["first_shards"] == comparison["second_shards"] == 2
+
+
+@pytest.mark.parametrize(
+    "bad_features, message",
+    [
+        ({"block4": torch.zeros(2, 2, 2)}, "ordered"),
+        (
+            {
+                "block4": torch.zeros(2, 2, 2),
+                "block8": torch.zeros(2, 2, 2),
+                "block12": torch.zeros(2, 2, 1),
+            },
+            "shape",
+        ),
+        (
+            {
+                "block4": torch.zeros(2, 2, 2),
+                "block8": torch.full((2, 2, 2), float("nan")),
+                "block12": torch.zeros(2, 2, 2),
+            },
+            "NaN or Inf",
+        ),
+    ],
+)
+def test_invalid_features_fail_before_commit(tmp_path, bad_features, message):
+    writer = FeatureCacheWriter(tmp_path, split="train2017", contract=contract(), target_shard_bytes=60)
+    with pytest.raises(ValueError, match=message):
+        writer.add(
+            sample_id="train2017/000000000001",
+            split="train2017",
+            image_path="images/train2017/000000000001.jpg",
+            image_sha256=image_sha(1),
+            features=bad_features,
+        )
+
+
+def test_corrupt_shard_fails_checksum_verification(tmp_path):
+    build_cache(tmp_path, values=(1,))
+    shard = next(tmp_path.glob("*.safetensors"))
+    with shard.open("r+b") as stream:
+        stream.seek(-1, 2)
+        byte = stream.read(1)
+        stream.seek(-1, 2)
+        stream.write(bytes([byte[0] ^ 0xFF]))
+
+    with pytest.raises(ValueError, match="shard checksum mismatch"):
+        verify_feature_cache(tmp_path)
+
+
+def test_failed_shard_save_preserves_part_file(tmp_path, monkeypatch):
+    safe_open, _save_file = cache_module._safetensors_api()
+
+    def fail_after_partial_write(_tensors, filename, metadata):
+        Path(filename).write_bytes(b"partial")
+        raise RuntimeError("simulated interruption")
+
+    monkeypatch.setattr(cache_module, "_safetensors_api", lambda: (safe_open, fail_after_partial_write))
+    writer = FeatureCacheWriter(tmp_path, split="train2017", contract=contract(), target_shard_bytes=60)
+    add_sample(writer, 1)
+
+    with pytest.raises(RuntimeError, match="simulated interruption"):
+        writer.flush()
+
+    assert (tmp_path / ".train2017-00000.safetensors.part").read_bytes() == b"partial"
+    assert not list(tmp_path.glob("*.safetensors"))
+
+
+def test_reader_reuses_and_evicts_shard_handles(tmp_path, monkeypatch):
+    build_cache(tmp_path)
+    safe_open, save_file = cache_module._safetensors_api()
+    calls = []
+
+    def counted_safe_open(*args, **kwargs):
+        calls.append(Path(args[0]).name)
+        return safe_open(*args, **kwargs)
+
+    monkeypatch.setattr(cache_module, "_safetensors_api", lambda: (counted_safe_open, save_file))
+    reader = FeatureCacheReader(tmp_path, max_open_shards=1)
+    reader.get("train2017/000000000001")
+    reader.get("train2017/000000000001")
+    reader.get("train2017/000000000002")
+    reader.get("train2017/000000000001")
+    reader.close()
+
+    assert calls == [
+        "train2017-00000.safetensors",
+        "train2017-00001.safetensors",
+        "train2017-00000.safetensors",
+    ]
+
+
+@pytest.mark.parametrize("value", (-1, 1.5, True))
+def test_reader_rejects_invalid_handle_cache_size(tmp_path, value):
+    with pytest.raises(ValueError, match="max_open_shards"):
+        FeatureCacheReader(tmp_path, max_open_shards=value)
+
 
 NAMES = ["block4", "block8", "block12"]
 
@@ -278,3 +551,49 @@ def test_real_npy_model_backward_and_checkpoint(tmp_path):
     restored = D1FoundationDetectionModel().to(device)
     initialize_mixture_loss_ema_buffer(restored)
     restored.load_state_dict(torch.load(state, map_location=device, weights_only=True), strict=True)
+
+
+def test_npy_conversion_preserves_source_and_rejects_corruption(tmp_path):
+    import torch
+    from scripts.d1.npy import convert_preserving_source
+    from ultralytics.nn.foundation.cache import FeatureCacheWriter
+    from ultralytics.nn.foundation.npy_cache import NpyFeatureCacheReader
+
+    split = "visdrone-train"
+    source, out = tmp_path / split, tmp_path / "npy"
+    contract = {
+        "model_id": "test",
+        "teacher_weights_sha256": "a" * 64,
+        "preprocessing_sha256": "b" * 64,
+        "output_layers": [4, 8, 12],
+        "feature_names": ["block4", "block8", "block12"],
+        "dtype": "float16",
+        "expected_shape": [384, 40, 40],
+    }
+    writer = FeatureCacheWriter(source, split=split, contract=contract, shard_prefix=split + "-r00")
+    feature = {
+        name: torch.full((384, 40, 40), float(i), dtype=torch.float16)
+        for i, name in enumerate(contract["feature_names"])
+    }
+    writer.add(
+        sample_id=split + "/original_id",
+        split=split,
+        image_path="images/" + split + "/original_id.jpg",
+        image_sha256="c" * 64,
+        features=feature,
+    )
+    writer.close()
+    before = {p.name: p.read_bytes() for p in source.iterdir() if p.is_file()}
+    first = convert_preserving_source(source, out)
+    assert first == convert_preserving_source(source, out)
+    assert before == {p.name: p.read_bytes() for p in source.iterdir() if p.is_file()}
+    reader = NpyFeatureCacheReader(out / split)
+    reader.verify_sample(split + "/original_id")
+    for name, value in reader.get(split + "/original_id").items():
+        assert torch.equal(value, feature[name])
+    path = out / split / "original_id.npy"
+    data = bytearray(path.read_bytes())
+    data[-1] ^= 1
+    path.write_bytes(data)
+    with pytest.raises(ValueError):
+        convert_preserving_source(source, out)
