+++ tests/test_d1_cache_cli.py
@@ -1,16 +1,13 @@
-"""Contracts for the D1 WP2 feature-cache command."""
+"""Deterministic extraction, resumable batches, and cache command validation."""
 
 from pathlib import Path
 from types import SimpleNamespace
-
 import cv2
 import numpy as np
 import pytest
 import torch
-
 from scripts.d1 import cache_features as cache
 from scripts.d1.cache_features import cache_contract, load_image, make_letterbox, split_paths
-
 
 REPO_ROOT = Path(__file__).resolve().parents[1]
